requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
gunicorn
psycopg2-binary==2.9.9
google-generativeai
PyJWT
twilio
